switched to using actual folders for generator templates instead of splitting on "-"
