+++ buffalo/cmd/dev.go
@@ -71,7 +71,10 @@ func startDevServer(ctx context.Context) error {
 	cfgFile := "./.buffalo.dev.yml"
 	_, err := os.Stat(cfgFile)
 	if err != nil {
-		g := rg.New()
+		g, err := rg.New()
+		if err != nil {
+			return err
+		}
 		err = g.Run("./", map[string]interface{}{
 			"name": "buffalo",
 		})
