+++ generators/newapp/templates/README.md.tmpl
