+++ generators/newapp/templates/actions/app.go.tmpl
