+++ generators/newapp/templates/actions/home.go.tmpl
