+++ generators/newapp/templates/actions/home_test.go.tmpl
