+++ generators/newapp/templates/actions/render.go.tmpl
