+++ generators/newapp/templates/dot-codeclimate.yml.tmpl
@@ -1,0 +1,20 @@
+engines:
+  fixme:
+    enabled: true
+  gofmt:
+    enabled: true
+  golint:
+    enabled: true
+  govet:
+    enabled: true
+exclude_paths:
+  - grifts/**/*
+  - "**/*_test.go"
+  - "*_test.go"
+  - "**_test.go"
+  - logs/*
+  - public/*
+  - templates/*
+ratings:
+  paths:
+    - "**.go"
