+++ generators/newapp/templates/dot-gitignore.tmpl
@@ -1,0 +1,11 @@
+vendor/
+**/*.log
+**/*.sqlite
+.idea/
+bin/
+tmp/
+node_modules/
+.sass-cache/
+rice-box.go
+public/assets/
+{{ name }}
