+++ generators/newapp/templates/grifts/routes.go.tmpl
