+++ generators/newapp/templates/main.go.tmpl
