+++ generators/newapp/templates/templates/application.html.tmpl
