+++ generators/newapp/templates/templates/index.html.tmpl
