+++ generators/refresh/refresh.go
@@ -1,37 +1,22 @@
 package refresh
 
-import "github.com/markbates/gentronics"
+import (
+	"github.com/gobuffalo/buffalo/generators/common"
+	"github.com/markbates/gentronics"
+)
 
 // New generator for a .buffalo.dev.yml file
-func New() *gentronics.Generator {
+func New() (*gentronics.Generator, error) {
 	g := gentronics.New()
 
-	g.Add(gentronics.NewFile(".buffalo.dev.yml", nRefresh))
-	return g
-}
+	files, err := common.Find("refresh")
+	if err != nil {
+		return nil, err
+	}
 
-const nRefresh = `app_root: .
-ignored_folders:
-- vendor
-- log
-- logs
-- assets
-- public
-- grifts
-- tmp
-- bin
-- node_modules
-- .sass-cache
-included_extensions:
-- .go
-- .html
-- .md
-- .js
-- .tmpl
-build_path: tmp
-build_delay: 200ns
-binary_name: {{name}}-build
-command_flags: []
-enable_colors: true
-log_name: buffalo
-`
+	for _, f := range files {
+		g.Add(gentronics.NewFile(f.WritePath, f.Body))
+	}
+
+	return g, nil
+}
